+++ src/Mock.php
@@ -17,7 +17,7 @@ final class Mock
      *
      * @var string
      */
-    private static string $defaultTag = '0.0.20';
+    private static string $defaultTag = '0.0.24';
 
     /**
      * The directory to download the mock to.